Make some big fixes. Refactor. Create slider. Rework files. Remove ProductFiles. Create LocalFileInterceptor. Complete orders.

+++ src/api/orders/orders.service.ts
@@ -5,16 +5,34 @@ import { Repository } from 'typeorm';
 import { Order } from '@/api/orders/entities/order.entity';
 import { UserService } from '@/api/user/user.service';
 import { CartService } from '@/api/cart/cart.service';
+import { OrderProduct } from '@/api/orders/entities/order-product.entity';
+import { CreateOrderProduct } from '@/api/orders/types/order-product';
 
 @Injectable()
 export class OrdersService {
   constructor(
-    @InjectRepository(Order) private readonly repository: Repository<Order>,
+    @InjectRepository(Order)
+    private readonly orderRepository: Repository<Order>,
+    @InjectRepository(OrderProduct)
+    private readonly orderProductRepository: Repository<OrderProduct>,
     @Inject(UserService) private readonly userService: UserService,
     @Inject(CartService) private readonly cartService: CartService,
   ) {}
 
-  async create(params: CreateOrderDto) {
+  public findAll(user_id: number): Promise<Order[]> {
+    return this.orderRepository.find({
+      where: { user: { id: user_id } },
+      relations: ['products', 'user'],
+    });
+  }
+
+  public createOrderProduct(params: CreateOrderProduct): Promise<OrderProduct> {
+    const newProduct = this.orderProductRepository.create(params);
+
+    return this.orderProductRepository.save(newProduct);
+  }
+
+  async create(params: CreateOrderDto): Promise<Order> {
     const user = await this.userService.findUser(params.user_id);
     const cart = await this.cartService.findAll(params.user_id);
 
@@ -26,17 +44,30 @@ export class OrdersService {
       throw new BadRequestException('cart is empty');
     }
 
-    const newOrder = this.repository.create({
+    const products = await Promise.all(
+      cart.items.map((item) =>
+        this.createOrderProduct({
+          ...item.product,
+          count: item.count,
+        }),
+      ),
+    );
+
+    const newOrder = this.orderRepository.create({
       ...params,
       user,
-      products: cart.items.map((cartItem) => cartItem.product),
+      products,
       status: 'Ожидает оплаты',
     });
 
-    await Promise.all(
-      cart.items.map((cartItem) => this.cartService.remove(cartItem.id)),
-    );
+    const savedOrder = await this.orderRepository.save(newOrder);
 
-    return this.repository.save(newOrder);
+    if (savedOrder) {
+      await Promise.all(
+        cart.items.map((cartItem) => this.cartService.remove(cartItem.id)),
+      );
+    }
+
+    return savedOrder;
   }
 }
